refactor: change folder in the playground
